added check to avoid centroid dropout
If a centroid has zero matches, its values will be
zeroed out (old approach). This reduces the total
number of centroids, and is greatly affected during
live video.
The new option checks for a centroid likely to be
dropped (has less matches than threshold). If below
threshold, randomly selects new values for the centroid.
Ideally, this should allow new matches as the items
in the frame change.